Adding save_files to generating_indxs.py

+++ src/scripts/generating_indxs.py
@@ -1,8 +1,8 @@
+import argparse
 import itertools
 import json
 import os.path
 import shutil
-
 import numpy as np
 from sklearn.utils import shuffle
 from imvc.ampute import Amputer
@@ -11,8 +11,14 @@ from imvc.impute import get_observed_view_indicator
 from settings import PROFILES_PATH, DATASET_TABLE_PATH, RANDOM_STATE, probs, amputation_mechanisms, runs_per_alg
 from src.utils import CommonOperations
 
-shutil.rmtree(PROFILES_PATH, ignore_errors=True)
-os.mkdir(PROFILES_PATH)
+parser = argparse.ArgumentParser()
+parser.add_argument('-continue_indxs', default=False, action='store_true')
+parser.add_argument('-save_results', default=False, action='store_true')
+args = parser.parse_args()
+
+if not args.continue_indxs:
+    shutil.rmtree(PROFILES_PATH, ignore_errors=True)
+    os.mkdir(PROFILES_PATH)
 
 datasets, two_view_datasets = CommonOperations.get_list_of_datasets(DATASET_TABLE_PATH)
 
@@ -20,6 +26,10 @@ for dataset_name in datasets:
     Xs, y, n_clusters = CommonOperations.load_Xs_y(dataset_name=dataset_name)
 
     for prob, amputation_mechanism, run_n in itertools.product(probs, amputation_mechanisms, runs_per_alg):
+        path = f"{dataset_name}_{prob}_{amputation_mechanism}_{run_n}.json"
+        path = os.path.join(PROFILES_PATH, path)
+        if os.path.exists(path):
+            continue
         try:
             random_state = RANDOM_STATE + run_n
             if (dataset_name in two_view_datasets) and (amputation_mechanism in ["MAR", "MNAR"]):
@@ -46,10 +56,9 @@ for dataset_name in datasets:
                 "error": str(exception),
             }
 
-        path = f"{dataset_name}_{prob}_{amputation_mechanism}_{run_n}.json"
-        path = os.path.join(PROFILES_PATH, path)
-        with open(path, "w") as f:
-            json.dump(dict_indxs, f)
+        if args.save_results:
+            with open(path, "w") as f:
+                json.dump(dict_indxs, f)
 
 print("Completed successfully!")
